update giay/* entity, repo, service, controller

+++ src/main/java/luckystore/datn/model/request/KichThuocRequest.java
@@ -24,12 +24,12 @@ public class KichThuocRequest {
     @NotNull(message = "Không được để trống chiều rộng")
     @Min(message = "Chiều rộng không được âm", value = 0)
     @Max(message = "Chiều rộng không hợp lệ, chiều rộng không được vượt quá 100", value = 100)
-    private Float chieuRong;
+    private Double chieuRong;
 
     @NotNull(message = "Không được để trống chiều dài")
     @Min(message = "Chiều dài không được âm", value = 0)
     @Max(message = "Chiều dài không hợp lệ, chiều dài không được vượt quá 100", value = 100)
-    private Float chieuDai;
+    private Double chieuDai;
 
     private String moTa;
